refactor(guide): 공식 가이드 보상 섹션을 OfficialRewardSection 으로 분리
RewardSection(+REWARD_POINTS·NAVER_GREEN)을 프레젠테이션 컴포넌트로 이동.
마크업/클래스 동일 — 시각 불변. 미사용 Info import 정리. 795→604 LOC (B).
